fix(Autocomplete): make DrawerList direction observer work
