+++ packages/dnb-eufemia/src/fragments/drawer-list/__tests__/DrawerList.test.js
@@ -14,39 +14,12 @@ import {
 } from '../../../core/jest/jestSetup'
 import Component from '../DrawerList'
 
-beforeAll(() => {
-  window.resizeTo = function resizeTo({
-    width = window.innerWidth,
-    height = window.innerHeight,
-  }) {
-    Object.assign(this, {
-      innerWidth: width,
-      innerHeight: height,
-    }).dispatchEvent(new this.Event('resize'))
+import {
+  mockImplementationForDirectionObserver,
+  testDirectionObserver,
+} from './DrawerListTestMocks'
 
-    // new setDirectionObserver implementation
-    jest
-      .spyOn(document.documentElement, 'clientWidth', 'get')
-      .mockImplementation(() => width)
-    jest
-      .spyOn(document.documentElement, 'clientHeight', 'get')
-      .mockImplementation(() => height)
-  }
-
-  window.scrollTo = function resizeTo({ top = window.pageYOffset }) {
-    Object.assign(this, {
-      pageYOffset: top,
-    }).dispatchEvent(new this.Event('scroll'))
-
-    // new setDirectionObserver implementation
-    jest
-      .spyOn(document.documentElement, 'scrollTop', 'get')
-      .mockImplementation(() => top)
-  }
-
-  // make sure we get the correct document.documentElement.clientHeight on startup
-  window.resizeTo({ height: window.innerHeight })
-})
+mockImplementationForDirectionObserver()
 
 const snapshotProps = {
   ...fakeProps(require.resolve('../DrawerList'), {
@@ -341,42 +314,7 @@ describe('DrawerList component', () => {
 
   it('has working direction observer', async () => {
     const Comp = mount(<Component {...props} data={mockData} />)
-    expect(Comp.props().direction).toBe('auto')
-
-    // the setDirectionObserver fn is changing this
-    expect(Comp.exists('.dnb-drawer-list--bottom')).toBe(true)
-    expect(
-      Comp.find('.dnb-drawer-list__options')
-        .instance()
-        .getAttribute('style')
-    ).toBe('max-height: 33.5rem;') // jsdom default is 768 innerHeight
-
-    window.resizeTo({
-      height: 640, // change innerHeight
-    })
-    await wait(100)
-
-    expect(Comp.exists('.dnb-drawer-list--bottom')).toBe(true)
-    expect(
-      Comp.find('.dnb-drawer-list__options')
-        .instance()
-        .getAttribute('style')
-    ).toBe('max-height: 28rem;')
-
-    window.scrollTo({
-      top: -640,
-    })
-    await wait(100)
-
-    // force re-render to get a updated state
-    Comp.update()
-
-    expect(Comp.exists('.dnb-drawer-list--top')).toBe(true)
-    expect(
-      Comp.find('.dnb-drawer-list__options')
-        .instance()
-        .getAttribute('style')
-    ).toBe('max-height: 28rem;') // is now min_height
+    await testDirectionObserver(Comp)
   })
 
   it('will call on_hide after "esc" key', () => {
@@ -516,4 +454,3 @@ const keydown = (Comp, keyCode) => {
   //   keyCode
   // })
 }
-const wait = (t) => new Promise((r) => setTimeout(r, t))
